Add simple Phonebook data structure
We are going to be trying to search through a list of contacts (which I
am calling a Phonebook for ease of discussion). We are going to store
these in some form; for now, the simplest possible answer is a hash
mapping Names to Numbers.

For the time being, we are going to consider phone numbers as strings;
this way we can accept arbitrary formatting (until we know better what
expectations Twilio has).