Teach Phonebook to translate words into dialcode
We want to be able to search from input given by a dialpad. This means
that we will want to index contacts by their dialpad equivalent ("Noah"
-> "6624") so that we can easily search our list.

To make this easy, here is a utility function: Phonebook.phoneword,
which maps a string to its digit equivalent.

Tests are implemented with RSpec (http://rspec.info/).