Add Phonebook#dialpad_search
Given a phonebook, we can now search it for name/number pairs based on
input from a dialpad. We look for exact matches that starting a first or
last name.

This is perhaps inefficient, in that we could precompute the search
index and do a direct lookup, but this is certainly the quick and dirty
way to code this feature.