Use the Directory number as outbound CallerID
We would like users to be able to identify when they are receiving a
directory call to help them distinguish it from telemarketers and other
spam (and/or to avoid work calls during off hours). Specifying the
directory number on outbound CallerID means that members of the
directory can add the directory number once and identify all future
calls.

Since we know that the user called the Directory, we can use the 'To'
parameter that Twilio passes us for this purpose. This way, we can avoid
adding configuration to a given server for its Twilio Number.